% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_results.R
\name{average_results}
\alias{average_results}
\title{Average results over valid days}
\usage{
average_results(data, minimum_wear_time = 10)
}
\arguments{
\item{data}{A dataframe obtained using the \code{\link{prepare_dataset}},  \code{\link{mark_wear_time}}, \code{\link{mark_intensity}} and \code{\link{recap_by_day}} functions.}

\item{minimum_wear_time}{A numeric value (in hours) to set the minimum wear time duration for validating a day.}
}
\value{
A dataframe.
}
\description{
This function computes, using valid days only, the mean of each of the metrics
obtained using the \code{\link{recap_by_day}} function.
}
\examples{
file <- system.file("extdata", "acc.agd", package = "activAnalyzer")
mydata <- prepare_dataset(data = file, epoch_len_out = 60, col_time_stamp = "timestamp")
mydata_with_wear_marks <- mark_wear_time(dataset = mydata)
mydata_with_intensity_marks <- mark_intensity(data = mydata_with_wear_marks, 
    equation = "Sasaki et al. (2011) [Adults]", sex = "male")
summary_by_day <- recap_by_day(data = mydata_with_intensity_marks, col_date = "date",  sex = "male")
average_results(data = summary_by_day)

}
