% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mx.R
\name{compute_mx}
\alias{compute_mx}
\title{Compute MX metric}
\usage{
compute_mx(x, n)
}
\arguments{
\item{x}{A vector of counts data.}

\item{n}{An integer setting the number of rows in correspondance with the
targetted amount of time.}
}
\value{
A numeric value.
}
\description{
An MX metric is the count/epoch value at and above which a given amount of the
most active minutes is spent (continuously or discontinuously).
}
