% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_peak_step_acc.R
\name{compute_peak_step_acc}
\alias{compute_peak_step_acc}
\title{Compute mean step accumulation (per min) from a given number of the best continous
or discontinuous minutes}
\usage{
compute_peak_step_acc(x, n)
}
\arguments{
\item{x}{A vector of steps data, each value corresponding to a total of steps
in a minute.}

\item{n}{An integer value setting the number of minutes to be used to compute
the metric.}
}
\value{
A numeric value.
}
\description{
Compute mean step accumulation (per min) from a given number of the best continous
or discontinuous minutes
}
