% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_pro_actigraph_score.R
\name{compute_pro_actigraph_score}
\alias{compute_pro_actigraph_score}
\title{Compute PROactive monitor-based physical activity score}
\usage{
compute_pro_actigraph_score(x, metric = c("steps", "vmu"))
}
\arguments{
\item{x}{A numeric value that should be the median of the daily steps counted during the measurement period (see Gimeno-Santos et al., 2015; doi: 10.1183/09031936.00183014).}

\item{metric}{A character value to indicate the metric for which the PROactive score should be obtained.}
}
\value{
A numeric value.
}
\description{
Compute PROactive monitor-based physical activity score
}
\examples{
compute_pro_actigraph_score(x= 3500, metric = "steps")

compute_pro_actigraph_score(x= 340, metric = "vmu")

}
