% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_radar.R
\name{coord_radar}
\alias{coord_radar}
\title{Set a new coordinate system for radar plots}
\usage{
coord_radar(theta = "x", start = 0, direction = 1)
}
\arguments{
\item{theta}{A character value.}

\item{start}{A numeric value.}

\item{direction}{A numeric value.}
}
\value{
A function setting a new coordinate system.
}
\description{
The code is from Erwan LE PENNEC, written on the following Stack overflow web page:
https://stackoverflow.com/questions/36579767/add-unit-labels-to-radar-plot-and-remove-outer-ring-ggplot2-spider-web-plot-co.
For information, because this piece of code was on Stack Overflow website, it is under CC 4.0 license, which is compatible
with the licence GPL v3 that is used for the {activAnalyzer} package (see: https://creativecommons.org/share-your-work/licensing-considerations/compatible-licenses/).
}
