% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fig_ratio_mpva_sed.R
\name{create_fig_ratio_mvpa_sed}
\alias{create_fig_ratio_mvpa_sed}
\title{Create a figure showing the mean daily MVPA/SED ratio}
\usage{
create_fig_ratio_mvpa_sed(score, language = c("en", "fr"))
}
\arguments{
\item{score}{A numeric value for daily MVPA/SED ratio.}

\item{language}{A character value for setting the language with which the figure should be created: "en" for english; "fr" for french.}
}
\value{
A ggplot object.
}
\description{
The function generates a figure showing the mean of the MVPA/SED ratio in correspondance with the Chastin et al. (2021; doi: 10.1123/jpah.2020-0635)
mortality hazard ratios.
}
\examples{
create_fig_ratio_mvpa_sed(score = 0.06)

}
