% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fig_steps.R
\name{create_fig_steps}
\alias{create_fig_steps}
\title{Create a figure showing the mean daily steps count}
\usage{
create_fig_steps(score, language = c("en", "fr"))
}
\arguments{
\item{score}{A numeric value for daily step count.}

\item{language}{A character value for setting the language with which the figure should be created: "en" for english; "fr" for french.}
}
\value{
A ggplot object.
}
\description{
The function generates a figure showing the mean of the daily step count in correspondance with the Tudor-Locke et al.
(2011; doi: 10.1186/1479-5868-8-79) categories.
}
\examples{
create_fig_steps(score = 12500)

}
