% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_all_analyses.R
\name{do_all_analyses}
\alias{do_all_analyses}
\title{Do all analyses at once}
\usage{
do_all_analyses(to_epoch = 60)
}
\arguments{
\item{to_epoch}{A numeric value to set the epoch required to collapse counts
in seconds.}
}
\value{
A dataset (1 row) with all computed metrics.
}
\description{
This function performs all analyses successively using the default data file associated to
the package. It is an internal function allowing the computation of the speed of the whole
analysis process, from the data importation to the final line of the results.
}
