% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_guidelines_status.R
\name{get_guidelines_status}
\alias{get_guidelines_status}
\title{Get WHO physical activity guidelines status}
\usage{
get_guidelines_status(value, language = c("en", "fr"))
}
\arguments{
\item{value}{A numeric value to indicate the daily mean of MET-hours spent at moderate-to-vigorous physical activity intensity.}

\item{language}{A character value for setting the language with which the table should be created: \code{en} for english; \code{fr} for french.}
}
\value{
A character string.
}
\description{
Get WHO physical activity guidelines status
}
\examples{
get_guidelines_status(value = 5)

}
