% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pa_period_info.R
\name{get_pa_period_info}
\alias{get_pa_period_info}
\title{Get relevant missing physical activity information indicated by the user of the app}
\usage{
get_pa_period_info(period)
}
\arguments{
\item{period}{A character string pointing to the module id from which information
have to be catched}
}
\value{
A dataframe
}
\description{
Get relevant missing physical activity information indicated by the user of the app
}
