% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pal_status.R
\name{get_pal_status}
\alias{get_pal_status}
\title{Get FAO physical activity level (PAL) status (http://www.fao.org/3/y5686e/y5686e07.htm#bm07.3)}
\usage{
get_pal_status(value, language = c("en", "fr"))
}
\arguments{
\item{value}{A numeric value to indicate the daily mean of PAL.}

\item{language}{A character value for setting the language with which the table should be created: "en" for english; "fr" for french.}
}
\value{
A character string.
}
\description{
Get FAO physical activity level (PAL) status (http://www.fao.org/3/y5686e/y5686e07.htm#bm07.3)
}
\examples{
get_pal_status(value = 1.8)
}
