% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ratio_mvpa_sed_comment.R
\name{get_ratio_mvpa_sed_comment}
\alias{get_ratio_mvpa_sed_comment}
\title{Get comment about the MPVA/SED ratio}
\usage{
get_ratio_mvpa_sed_comment(value, language = c("en", "fr"))
}
\arguments{
\item{value}{A numeric value to indicate the daily mean of MVPA/SED ratio.}

\item{language}{A character value for setting the language with which the table should be created: \code{en} for english; \code{fr} for french.}
}
\value{
A character string.
}
\description{
Get comment about the MPVA/SED ratio
}
\examples{
get_ratio_mvpa_sed_comment(value = 0.03)
}
