% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{prepare_dataset}
\alias{prepare_dataset}
\title{Prepare accelerometer data}
\usage{
prepare_dataset(data, epoch_len_out = 60, col_time_stamp = "timestamp")
}
\arguments{
\item{data}{Path to an .agd file that was exported from Actilife software.}

\item{epoch_len_out}{A numeric value (in seconds) setting the length of the desired epoch for accumulating accelerometer data. For now, only 60 can be used.}

\item{col_time_stamp}{A character value that corresponds to the name of the variable that contains date and time information.}
}
\value{
A dataframe.
}
\description{
This function performs the following actions: it reads the .agd file, then it collapses
accelerometer data based on the configured epoch length (for now, only 60 s can be used),
then it creates the vector magnitude variable, creates \code{date} and \code{time} columns, and finally organizes columns in a given order.

The .agd file must contain at least the following columns:
\itemize{
\item \strong{axis1}
\item \strong{axis2}
\item \strong{axis3}
\item \strong{steps}
\item \strong{inclinestanding}
\item \strong{inclinesitting}
\item \strong{inclinelying}
}
}
\examples{
file <- system.file("extdata", "acc.agd", package = "activAnalyzer")
prepare_dataset(data = file, epoch_len_out = 60, col_time_stamp = "timestamp")

}
