% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recap_by_day.R
\name{recap_by_day}
\alias{recap_by_day}
\title{Summarize results by day}
\usage{
recap_by_day(
  data,
  col_date = "date",
  col_time = "time",
  valid_wear_time_start = "00:00:00",
  valid_wear_time_end = "23:59:00",
  age = 40,
  weight = 70,
  sex = c("male", "female", "undefined")
)
}
\arguments{
\item{data}{A dataframe obtained using the \code{\link{prepare_dataset}}, \code{\link{mark_wear_time}}, and then the \code{\link{mark_intensity}} functions.}

\item{col_date}{A character value to indicate the name of the date variable.}

\item{col_time}{A character value indicating the name of the variable where time information is provided.}

\item{valid_wear_time_start}{A character value with the HH:MM:SS format to set the start of the daily period to consider for computing valid wear time.}

\item{valid_wear_time_end}{A character value with the HH:MM:SS format to set the end of the daily period to consider for computing valid wear time.}

\item{age}{A numeric value in yr.}

\item{weight}{A numeric value in kg.}

\item{sex}{A character value.}
}
\value{
A dataframe.
}
\description{
This function summarizes accelerometer results for each day of the measurement period.
}
\details{
The following metrics are computed:
\itemize{
\item \strong{wear_time:} total wear time computed using the daily period defined in the function
\item \strong{total_counts_axis1:} total counts for the vertical axis
\item \strong{total_counts_vm:} total counts for the vector magnitude
\item \strong{axis1_per_min:} mean of the counts per minute for the vertical axis (during the considered wear time only)
\item \strong{vm_per_min:} mean of the counts per minute for the vector magnitude (during the considered wear time only)
\item \strong{total_steps:} total step count
\item \strong{total_kcal:} total kilocalories
\item \strong{minutes_SED:} total minutes spent in SED behavior
\item \strong{minutes_LPA:} total minutes spent in LPA behavior
\item \strong{minutes_MPA:} total minutes spent in MPA behavior
\item \strong{minutes_VPA:} total minutes spent in VPA behavior
\item \strong{minutes_MVPA:} total minutes spent in MVPA behavior
\item \strong{percent_SED:} proportion of wear time (over the 24-h cycle) spent in SED behavior
\item \strong{percent_LPA:} proportion of wear time (over the 24-h cycle) spent in LPA behavior
\item \strong{percent_MPA:} proportion of wear time (over the 24-h cycle) spent in MPA behavior
\item \strong{percent_VPA:} proportion of wear time (over the 24-h cycle) spent in VPA behavior
\item \strong{percent_MVPA:} proportion of wear time (over the 24-h cycle) spent in MPVA behavior
\item \strong{max_steps_60min:} best step accumulation per minute averaged over a window of 60 continuous minutes
\item \strong{max_steps_30min:} best step accumulation per minute averaged over a window of 30 continuous minutes
\item \strong{max_steps_20min:} best step accumulation per minute averaged over a window of 20 continuous minutes
\item \strong{max_steps_5min:} best step accumulation per minute averaged over a window of 5 continuous minutes
\item \strong{max_steps_1min:} best step accumulation per minute over a window of 1 minute
\item \strong{peak_steps_60min:} step accumulation per minute averaged over the best 60 continuous or discontinuous minutes
\item \strong{peak_steps_30min:} step accumulation per minute averaged over the best 30 continuous or discontinuous minutes
\item \strong{peak_steps_20min:} step accumulation per minute averaged over the best 20 continuous or discontinuous minutes
\item \strong{peak_steps_5min:} step accumulation per minute averaged over the best 5 continuous or discontinuous minutes
\item \strong{peak_steps_1min:} step accumulation per minute over the best minute (same result as for \code{max_steps_1min})
\item \strong{mets_hours_mvpa:} total MET-hours spent during MPVA behavior
\item \strong{ratio_mvpa_sed:} ratio between MVPA and SED times (\code{minutes_MVPA} / \code{minutes_SED})
}

PAL is computed by dividing total energy expenditure (TEE) by BMR. TEE is obtained by summing
the kilocalories measured during wear time epochs and the kilocalories related to BMR expended during nonwear time
epochs (it is assumed that the periods where the device was not worn corresponded to
sleeping periods, during which energy expenditure is near of BMR), and by multiplying this sum
by 10/9 to take into account the thermic effect of food. Of course, such calculations may
conduct to underestimate TEE and PAL if the device was removed during prolonged periods of
physical activity. Moreover, even if the device was correctly worn, the estimate of PAL is very
approximate since both BMR and kilocalories are estimated using methods
that may not be accurate at the individual level.
}
\examples{
file <- system.file("extdata", "acc.agd", package = "activAnalyzer")
mydata <- prepare_dataset(data = file, epoch_len_out = 60, col_time_stamp = "timestamp")
mydata_with_wear_marks <- mark_wear_time(
    dataset = mydata, 
    TS = "timestamp", 
    cts  = "vm", 
    frame = 90, 
    allowanceFrame = 2, 
    streamFrame = 30
    )
mydata_with_intensity_marks <- mark_intensity(
    data = mydata_with_wear_marks, 
    col_axis = "vm", 
    equation = "Sasaki et al. (2011) [Adults]",
    sed_cutpoint = 200, 
    mpa_cutpoint = 2690, 
    vpa_cutpoint = 6167, 
    age = 32,
    weight = 67,
    sex = "male",
    col_steps = "steps"
    )
recap_by_day(
    data = mydata_with_intensity_marks, 
    age = 32, 
    weight = 67, 
    sex = "male",
    valid_wear_time_start = "07:00:00",
    valid_wear_time_end = "22:00:00"
    )

}
