% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_with_intensity.R
\name{plot_data_with_intensity}
\alias{plot_data_with_intensity}
\title{Plot accelerometer data for each day with both nonwear time and physical activity intensity categories}
\usage{
plot_data_with_intensity(
  data,
  metric = "axis1",
  col_time = "time",
  col_nonwear = "non_wearing_count",
  col_wear = "wearing_count",
  col_date = "date"
)
}
\arguments{
\item{data}{A dataframe obtained using the \code{\link{mark_wear_time}} and then the \code{\link{mark_intensity}} functions.}

\item{metric}{A character value to indicate the name of the variable to be plotted against time.}

\item{col_time}{A character value to indicate the name of the variable to plot time data.}

\item{col_nonwear}{A character value to indicate the name of the variable used to count nonwear time.}

\item{col_wear}{A character value to indicate the name of the variable used to count wear time.}

\item{col_date}{A character value to indicate the name of the variable to plot date data.}
}
\value{
A \code{ggplot} object.
}
\description{
This function plots accelerometer data with intensity categories against time for each day of measurement,
with the possibility to specify the metric to visualize.
}
\examples{
file <- system.file("extdata", "acc.agd", package = "activAnalyzer")
mydata <- prepare_dataset(data = file, epoch_len_out = 60, col_time_stamp = "timestamp")
mydata_with_wear_marks <- mark_wear_time(
    dataset = mydata, 
    TS = "timestamp", 
    cts  = "vm", 
    frame = 90, 
    allowanceFrame = 2, 
    streamFrame = 30
    )
mydata_with_intensity_marks <- mark_intensity(
    data = mydata_with_wear_marks, 
    col_axis = "vm", 
    equation = "Sasaki et al. (2011) [Adults]",
    sed_cutpoint = 200, 
    mpa_cutpoint = 2690, 
    vpa_cutpoint = 6167, 
    age = 32,
    weight = 67,
    sex = "male",
    col_steps = "steps"
    )
plot_data_with_intensity(
    data = mydata_with_intensity_marks,
    metric = "vm"
    )

}
