% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_dataset.R
\name{prepare_dataset}
\alias{prepare_dataset}
\title{Prepare accelerometer data}
\usage{
prepare_dataset(data)
}
\arguments{
\item{data}{Path to an .agd file that was exported from Actilife software.}
}
\value{
A dataframe.
}
\description{
This function reads an .agd file. The .agd file must contain at least the following columns:
\itemize{
\item \strong{axis1}
\item \strong{axis2}
\item \strong{axis3}
\item \strong{steps}
\item \strong{inclineStanding}
\item \strong{inclineSitting}
\item \strong{inclineLying}
}
}
\examples{
file <- system.file("extdata", "acc.agd", package = "activAnalyzer")
mydata <- prepare_dataset(data = file)
head(mydata)

}
