% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recap_by_day.R
\name{recap_by_day}
\alias{recap_by_day}
\title{Summarize results by day}
\usage{
recap_by_day(
  data,
  col_date = "date",
  age = 40,
  weight = 70,
  sex = c("male", "female", "undefined")
)
}
\arguments{
\item{data}{A dataframe obtained using the \code{\link{prepare_dataset}}, \code{\link{mark_wear_time}}, and \code{\link{mark_intensity}} functions.}

\item{col_date}{A character value to indicate the name of the date variable.}

\item{age}{A numeric value in yr.}

\item{weight}{A numeric value in kg.}

\item{sex}{A character value.}
}
\value{
A dataframe.
}
\description{
This function summarizes results for each day of the measurement period.
}
\details{
The following metrics are computed:
\itemize{
\item \strong{wear_time:} total wear time
\item \strong{total_counts_axis1}: total counts for the vertical axis
\item \strong{total_counts_vm:} total counts for the vector magnitude
\item \strong{total_steps:} total step count
\item \strong{total_kcal_wear_time:} total kilocalories
\item \strong{minutes_SED:} total minutes spent in SED behavior
\item \strong{minutes_LPA:} total minutes spent in LPA behavior
\item \strong{minutes_MPA:} total minutes spent in MPA behavior
\item \strong{minutes_VPA:} total minutes spent in VPA behavior
\item \strong{minutes_MVPA:} total minutes spent in MVPA behavior
\item \strong{percent_SED:} proportion of wear time spent in SED behavior
\item \strong{percent_LPA:} proportion of wear time spent in LPA behavior
\item \strong{percent_MPA:} proportion of wear time spent in MPA behavior
\item \strong{percent_VPA:} proportion of wear time spent in VPA behavior
\item \strong{percent_MVPA:} proportion of wear time spent in MPVA behavior
\item \strong{max_steps_60min:} maximum average step accumulation per minute over a window of 60 continuous minutes
\item \strong{max_steps_30min:} maximum average step accumulation per minute over a window of 30 continuous minutes
\item \strong{max_steps_20min:} maximum average step accumulation per minute over a window of 20 continuous minutes
\item \strong{max_steps_5min:} maximum average step accumulation per minute over a window of 5 continuous minutes
\item \strong{max_steps_1min:} maximum average step accumulation per minute over a window of 1 minute
\item \strong{peak_steps_60min:} maximum average step accumulation per minute over the best 60 continuous or discontinuous minutes
\item \strong{peak_steps_30min:} maximum average step accumulation per minute over the best 30 continuous or discontinuous minutes
\item \strong{peak_steps_20min:} maximum average step accumulation per minute over the best 20 continuous or discontinuous minutes
\item \strong{peak_steps_5min:} maximum average step accumulation per minute over the best 5 continuous or discontinuous minutes
\item \strong{peak_steps_1min:} maximum average step accumulation per minute over the best 1 minute (same result as for \code{Max_steps_1min})
\item \strong{mets_hours_mvpa:} total MET-hours spent during MPVA behavior
\item \strong{ratio_mvpa_sed:} ratio between MVPA and SED times (minutes_MVPA / minutes_SED)
}

PAL is computed as follows: total energy expenditure (TEE) is divided by BMR. TEE is obtained by summing
the kilocalories measured during wear time epochs and the kilocalories likely expended during nonwear time
epochs (that is, kilocalories associated to BMR, as it is assumed that the device was mainly not worn during
sleeping periods if any, periods during which energy expenditure is near of BMR), and by multiplying this sum
by 10/9 to take into account the thermic effect of food. Of course, such calculations may
conduct to underestimate TEE and PAL if the device was removed during prolonged periods of
physical activity. Moreover, even if the device was correctly worn, the estimate of PAL is very
approximate since both BMR and kilocalories related to wear time are estimated using methods
that may not be accurate at the individual level.
}
\examples{
file <- system.file("extdata", "acc.agd", package = "activAnalyzer")
mydata <- prepare_dataset(data = file, epoch_len_out = 60, col_time_stamp = "timestamp")
mydata_with_wear_marks <- mark_wear_time(dataset = mydata)
mydata_with_intensity_marks <- mark_intensity(data = mydata_with_wear_marks, 
    equation = "Sasaki et al. (2011) [Adults]", age = 32, weight = 67, sex = "male")
recap_by_day(data = mydata_with_intensity_marks, col_date = "date", sex = "male")

}
